splitting v0 tests for purchaseKey
